Get ASV suite up and running
